refactor: move security settings router to routers
